requirements.txt
fastapi
uvicorn[standard]
pypdf2
python-docx
transformers
langchain
torch
python-multipart
fastapi-cors
pydantic
huggingface_hub
python-dotenv
langchain_community
groq
nltk
spacy
langchain-huggingface
nltk
python-dotenv
TTS
pyttsx3
pydub
feedgen
streamlit
